requirements.txt
matplotlib==2.2.2
requests==2.23.0
clint==0.5.1
numpy==1.17.4
tensorflow==1.14.0
click==7.1.1
Pillow==7.0.0
scikit_learn==0.22.2.post1
